Added move sequence comparison function

+++ test/033_inverse_move/inverse_move_tests.c
@@ -3,11 +3,11 @@
 extern char *movestr[];
 
 int64_t readmoves(const char *, size_t n, size_t m,
-    uint64_t *, uint64_t *, uint8_t [n], uint8_t [m]);
+    size_t *, size_t *, uint8_t [n], uint8_t [m]);
 uint8_t inverse_move(uint8_t);
 
 void run(void) {
-	uint64_t inv, nor;
+	size_t inv, nor;
 	int64_t tot;
 	uint8_t moves[2];
