+++ test/062_transform_move/transform_move_tests.c
@@ -5,7 +5,7 @@
 cube_t applytrans(cube_t, const char *);
 uint8_t transform_move(uint8_t, uint8_t);
 int64_t readmoves(const char *, size_t n, size_t m,
-    uint64_t *, uint64_t *, uint8_t [n], uint8_t [m]);
+    size_t *, size_t *, uint8_t [n], uint8_t [m]);
 oriented_cube_t move_extended(oriented_cube_t, uint8_t);
 oriented_cube_t applymoves(oriented_cube_t, const char *);
 uint8_t readtrans(const char[static NISSY_SIZE_TRANSFORMATION]);
@@ -13,7 +13,7 @@ uint8_t readtrans(const char[static NISSY_SIZE_TRANSFORMATION]);
 void run(void) {
 	char movestr[STRLENMAX], transtr[STRLENMAX], cubestr[STRLENMAX];
 	uint8_t t, moves[MAXMOVES];
-	uint64_t i, n, ninv;
+	size_t i, n, ninv;
 	int64_t tot;
 	oriented_cube_t cube;
